Added auth restrictions for different api endpoints and imported users/me endpoint from auth/router to main.py

+++ backend/main.py
@@ -1,9 +1,11 @@
+from typing import Annotated
 from fastapi import Depends, FastAPI, HTTPException
 from sqlalchemy.orm import Session
 
 from . import crud, models, schemas
 from .database import SessionLocal, engine
 from .auth.router import router as auth_router
+from .auth.dependencies import get_current_user
 
 #Creates the db tables
 models.Base.metadata.create_all(bind=engine)
@@ -33,38 +35,50 @@ def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
     users = crud.get_users(db, skip = skip, limit = limit)
     return users
 
-@app.get("/users/{user_id}", response_model=schemas.UserInDB, tags=["users"])
-def read_user(user_id: int, db: Session = Depends(get_db)):
-    db_user = crud.get_user_by_id(db = db, user_id = user_id)
-    if db_user is None:
-        raise HTTPException(status_code=404, detail="User not found")
-    return db_user
+@app.get("/users/me/", response_model=schemas.UserInDB, tags=["users"])
+async def read_user(current_user: Annotated[schemas.UserInDB, Depends(get_current_user)]):
+    return current_user
+
 
 @app.delete("/users/{user_id}", response_model=dict, tags=["users"])
-def delete_user(user_id: int, db: Session = Depends(get_db)):
+def delete_user(user_id: int, db: Session = Depends(get_db), current_user: schemas.UserInDB = Depends(get_current_user)):
+    if current_user.id != user_id:
+        raise HTTPException(status_code=403, detail="Not authorized to delete this user")
     return crud.delete_user(db = db, user_id = user_id)
 
 # ----- Path Operations for Tasks below -----
 
-@app.post("/users/{user_id}/tasks/", response_model=schemas.TaskInDB, tags=["tasks"])
-def create_task_for_user(user_id: int, task: schemas.TaskCreate, db: Session = Depends(get_db)):
-    return crud.create_task(db = db, task = task, user_id = user_id)
+@app.post("/tasks/", response_model=schemas.TaskInDB, tags=["tasks"])
+def create_task_for_user(task: schemas.TaskCreate, db: Session = Depends(get_db), current_user: schemas.UserInDB = Depends(get_current_user)):
+    return crud.create_task(db = db, task = task, user_id = current_user.id)
 
-@app.get("/users/{user_id}/tasks/", response_model=list[schemas.TaskInDB], tags=["tasks"])
-def read_tasks(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    return crud.get_tasks(db, skip = skip, limit = limit)
+@app.get("/tasks/", response_model=list[schemas.TaskInDB], tags=["tasks"])
+def read_tasks(skip: int = 0, limit: int = 100, db: Session = Depends(get_db), current_user: schemas.UserInDB = Depends(get_current_user)):
+    return crud.get_tasks(db, user_id = current_user.id, skip = skip, limit = limit)
 
-@app.get("/users/{user_id}/tasks/{task_id}", response_model=schemas.TaskInDB, tags=["tasks"])
-def read_task(task_id: int, db: Session = Depends(get_db)):
-    db_task = crud.get_task_by_id(db = db, task_id = task_id)
+@app.get("/tasks/{task_id}", response_model=schemas.TaskInDB, tags=["tasks"])
+def read_task(task_id: int, db: Session = Depends(get_db), current_user: schemas.UserInDB = Depends(get_current_user)):
+    db_task = crud.get_task_by_id(db=db, task_id=task_id)
     if db_task is None:
-        raise HTTPException(status_code=404, detail="Task not found")
+        raise HTTPException(status_code=404, detail="Task doesn't exist")
+    elif db_task.owner_id != current_user.id:
+        raise HTTPException(status_code=404, detail="Not authorized to view this task")
     return db_task
 
-@app.put("/users/{user_id}/tasks/{task_id}", response_model=schemas.TaskInDB, tags=["tasks"])
-def update_task(task_id: int, task: schemas.TaskUpdate, db: Session = Depends(get_db)):
+@app.put("/tasks/{task_id}", response_model=schemas.TaskInDB, tags=["tasks"])
+def update_task(task_id: int, task: schemas.TaskUpdate, db: Session = Depends(get_db), current_user: schemas.UserInDB = Depends(get_current_user)):
+    db_task = crud.get_task_by_id(db=db, task_id=task_id)
+    if db_task is None:
+        raise HTTPException(status_code=404, detail="Task doesn't exist")
+    elif db_task.owner_id != current_user.id:
+        raise HTTPException(status_code=404, detail="Not authorized to update this task")
     return crud.update_task(db = db, task = task, task_id = task_id)
 
-@app.delete("/users/{user_id}/tasks/{task_id}", response_model=dict, tags=["tasks"])
-def delete_task(task_id: int, db: Session = Depends(get_db)):
+@app.delete("/tasks/{task_id}", response_model=dict, tags=["tasks"])
+def delete_task(task_id: int, db: Session = Depends(get_db), current_user: schemas.UserInDB = Depends(get_current_user)):
+    db_task = crud.get_task_by_id(db=db, task_id=task_id)
+    if db_task is None:
+        raise HTTPException(status_code=404, detail="Task doesn't exist")
+    elif db_task.owner_id != current_user.id:
+        raise HTTPException(status_code=404, detail="Not authorized to delete this task")
     return crud.delete_task(db = db, task_id = task_id)
